Working example of connected component typechecking
Notice how this will ONLY work if the container component
is being imported and used in another file (or if it was used
in the same file it would also work).

If all you have is the export of the container component,
flow doesn't complain about errors.

To test this, comment out everything in the `index.js` file
and run flow.